extract classification description func from config
