+++ dev/services/wms/ows_refactored/land_and_vegetation/landcover/ows_c2_lc_cyear_cfg.py
@@ -1,4 +1,6 @@
 from ows_refactored.ows_reslim_cfg import reslim_wms_min_zoom_15_cache_rules
+from ows_refactored.land_and_vegetation.landcover.lc_class_desc import class_description
+
 
 bands_landcover = {
     "level3": [],
@@ -263,108 +265,6 @@ style_c2_level4 = {
     },
 }
 
-level3_descriptions = {
-    0: "No data",
-    111: "Cultivated Terrestrial Vegetation",
-    112: "Natural Terrestrial Vegetation",
-    123: "Cultivated Aquatic Vegetation",
-    124: "Natural Aquatic Vegetation",
-    215: "Artificial Surface",
-    216: "Natural Bare Surface",
-    220: "Water"
-}
-
-level4_descriptions = {
-    0: "No data",
-    1: 'Cultivated Terrestrial Vegetated:',
-    3: 'Cultivated Terrestrial Vegetated: Herbaceous',
-    4: 'Cultivated Terrestrial Vegetated: Closed (> 65 %)',
-    5: 'Cultivated Terrestrial Vegetated: Open (40 to 65 %)',
-    6: 'Cultivated Terrestrial Vegetated: Open (15 to 40 %)',
-    7: 'Cultivated Terrestrial Vegetated: Sparse (4 to 15 %)',
-    8: 'Cultivated Terrestrial Vegetated: Scattered (1 to 4 %)',
-    14: 'Cultivated Terrestrial Vegetated: Herbaceous Closed (> 65 %)',
-    15: 'Cultivated Terrestrial Vegetated: Herbaceous Open (40 to 65 %)',
-    16: 'Cultivated Terrestrial Vegetated: Herbaceous Open (15 to 40 %)',
-    17: 'Cultivated Terrestrial Vegetated: Herbaceous Sparse (4 to 15 %)',
-    18: 'Cultivated Terrestrial Vegetated: Herbaceous Scattered (1 to 4 %)',
-    19: 'Natural Terrestrial Vegetated:',
-    20: 'Natural Terrestrial Vegetated: Woody',
-    21: 'Natural Terrestrial Vegetated: Herbaceous',
-    22: 'Natural Terrestrial Vegetated: Closed (> 65 %)',
-    23: 'Natural Terrestrial Vegetated: Open (40 to 65 %)',
-    24: 'Natural Terrestrial Vegetated: Open (15 to 40 %)',
-    25: 'Natural Terrestrial Vegetated: Sparse (4 to 15 %)',
-    26: 'Natural Terrestrial Vegetated: Scattered (1 to 4 %)',
-    27: 'Natural Terrestrial Vegetated: Woody Closed (> 65 %)',
-    28: 'Natural Terrestrial Vegetated: Woody Open (40 to 65 %)',
-    29: 'Natural Terrestrial Vegetated: Woody Open (15 to 40 %)',
-    30: 'Natural Terrestrial Vegetated: Woody Sparse (4 to 15 %)',
-    31: 'Natural Terrestrial Vegetated: Woody Scattered (1 to 4 %)',
-    32: 'Natural Terrestrial Vegetated: Herbaceous Closed (> 65 %)',
-    33: 'Natural Terrestrial Vegetated: Herbaceous Open (40 to 65 %)',
-    34: 'Natural Terrestrial Vegetated: Herbaceous Open (15 to 40 %)',
-    35: 'Natural Terrestrial Vegetated: Herbaceous Sparse (4 to 15 %)',
-    36: 'Natural Terrestrial Vegetated: Herbaceous Scattered (1 to 4 %)',
-    55: 'Natural Aquatic Vegetated:',
-    56: 'Natural Aquatic Vegetated: Woody',
-    57: 'Natural Aquatic Vegetated: Herbaceous',
-    58: 'Natural Aquatic Vegetated: Closed (> 65 %)',
-    59: 'Natural Aquatic Vegetated: Open (40 to 65 %)',
-    60: 'Natural Aquatic Vegetated: Open (15 to 40 %)',
-    61: 'Natural Aquatic Vegetated: Sparse (4 to 15 %)',
-    62: 'Natural Aquatic Vegetated: Scattered (1 to 4 %)',
-    63: 'Natural Aquatic Vegetated: Woody Closed (> 65 %)',
-    64: 'Natural Aquatic Vegetated: Woody Closed (> 65 %) Water > 3 months (semi-) permanent',
-    65: 'Natural Aquatic Vegetated: Woody Closed (> 65 %) Water < 3 months (temporary or seasonal)',
-    66: 'Natural Aquatic Vegetated: Woody Open (40 to 65 %)',
-    67: 'Natural Aquatic Vegetated: Woody Open (40 to 65 %) Water > 3 months (semi-) permanent',
-    68: 'Natural Aquatic Vegetated: Woody Open (40 to 65 %) Water < 3 months (temporary or seasonal)',
-    69: 'Natural Aquatic Vegetated: Woody Open (15 to 40 %)',
-    70: 'Natural Aquatic Vegetated: Woody Open (15 to 40 %) Water > 3 months (semi-) permanent',
-    71: 'Natural Aquatic Vegetated: Woody Open (15 to 40 %) Water < 3 months (temporary or seasonal)',
-    72: 'Natural Aquatic Vegetated: Woody Sparse (4 to 15 %)',
-    73: 'Natural Aquatic Vegetated: Woody Sparse (4 to 15 %) Water > 3 months (semi-) permanent',
-    74: 'Natural Aquatic Vegetated: Woody Sparse (4 to 15 %) Water < 3 months (temporary or seasonal)',
-    75: 'Natural Aquatic Vegetated: Woody Scattered (1 to 4 %)',
-    76: 'Natural Aquatic Vegetated: Woody Scattered (1 to 4 %) Water > 3 months (semi-) permanent',
-    77: 'Natural Aquatic Vegetated: Woody Scattered (1 to 4 %) Water < 3 months (temporary or seasonal)',
-    78: 'Natural Aquatic Vegetated: Herbaceous Closed (> 65 %)',
-    79: 'Natural Aquatic Vegetated: Herbaceous Closed (> 65 %) Water > 3 months (semi-) permanent',
-    80: 'Natural Aquatic Vegetated: Herbaceous Closed (> 65 %) Water < 3 months (temporary or seasonal)',
-    81: 'Natural Aquatic Vegetated: Herbaceous Open (40 to 65 %)',
-    82: 'Natural Aquatic Vegetated: Herbaceous Open (40 to 65 %) Water > 3 months (semi-) permanent',
-    83: 'Natural Aquatic Vegetated: Herbaceous Open (40 to 65 %) Water < 3 months (temporary or seasonal)',
-    84: 'Natural Aquatic Vegetated: Herbaceous Open (15 to 40 %)',
-    85: 'Natural Aquatic Vegetated: Herbaceous Open (15 to 40 %) Water > 3 months (semi-) permanent',
-    86: 'Natural Aquatic Vegetated: Herbaceous Open (15 to 40 %) Water < 3 months (temporary or seasonal)',
-    87: 'Natural Aquatic Vegetated: Herbaceous Sparse (4 to 15 %)',
-    88: 'Natural Aquatic Vegetated: Herbaceous Sparse (4 to 15 %) Water > 3 months (semi-) permanent',
-    89: 'Natural Aquatic Vegetated: Herbaceous Sparse (4 to 15 %) Water < 3 months (temporary or seasonal)',
-    90: 'Natural Aquatic Vegetated: Herbaceous Scattered (1 to 4 %)',
-    91: 'Natural Aquatic Vegetated: Herbaceous Scattered (1 to 4 %) Water > 3 months (semi-) permanent',
-    92: 'Natural Aquatic Vegetated: Herbaceous Scattered (1 to 4 %) Water < 3 months (temporary or seasonal)',
-    93: 'Artificial Surface',
-    94: 'Natural Surface',
-    95: 'Natural Surface: Sparsely vegetated',
-    96: 'Natural Surface: Very sparsely vegetated',
-    97: 'Natural Surface: Bare areas, unvegetated',
-    98: 'Water',
-    99: 'Water',
-    100: 'Water: Tidal area',
-    101: 'Water: Perennial (> 9 months)',
-    102: 'Water: Non-perennial (7 to 9 months)',
-    103: 'Water: Non-perennial (4 to 6 months)',
-    104: 'Water: Non-perennial (1 to 3 months)',
-}
-
-
-def class_description(data):
-    return {
-        "level3_desc": level3_descriptions[data["level3"]],
-        "level4_desc": level4_descriptions[data["level4"]],
-    }
-
 
 layers = {
     "title": "DEA Land Cover (Landsat)",
@@ -441,10 +341,6 @@ layers = {
                 "include_custom": {
                     "description": {
                         "function": class_description,
-                        "pass_product_cfg": False,
-                        "kwargs": {
-                            "template": ""
-                        }
                     }
                 }
             }
